refactor: clean up unused imports in audit and user test files
