+++ src/test/kotlin/data/datasource/mapper/UserMapperTest.kt
@@ -2,10 +2,8 @@ package data.datasource.mapper
 
 import com.google.common.truth.Truth.assertThat
 import org.baghdad.data.datasource.mapper.user.UserMapper
-import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.enums.UserType
 import org.baghdad.data.dto.UserDto
-import org.baghdad.logic.model.entities.UserType
+import org.baghdad.logic.model.enums.UserType
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
 import java.util.*
