+++ src/test/kotlin/data/local/UserDataSourceTest.kt
@@ -7,14 +7,12 @@ import kotlinx.coroutines.test.runTest
 import org.baghdad.data.datasource.DataSource
 import org.baghdad.data.dto.UserDto
 import org.baghdad.data.local.UserDataSource
-import org.baghdad.logic.model.entities.UserEntity
+import org.baghdad.logic.model.enums.UserType
 import org.baghdad.logic.model.exceptions.UserNotFoundException
-import org.baghdad.logic.model.entities.UserType
-import org.baghdad.logic.model.exceptions.user.UserNotFoundException
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import java.util.UUID
+import java.util.*
 
 class UserDataSourceTest {
